Refactor the spreaded function into module
